+++ controller/api/util/api_utils.go
@@ -21,7 +21,8 @@ import (
 */
 
 var (
-	defaultMetricTimeWindow = "1m"
+	defaultMetricTimeWindow    = "1m"
+	metricTimeWindowLowerBound = time.Second * 15 //the window value needs to equal or larger than that
 
 	// ValidTargets specifies resource types allowed as a target:
 	// target resource on an inbound query
@@ -145,10 +146,15 @@ func GRPCError(err error) error {
 func BuildStatSummaryRequest(p StatsSummaryRequestParams) (*pb.StatSummaryRequest, error) {
 	window := defaultMetricTimeWindow
 	if p.TimeWindow != "" {
-		_, err := time.ParseDuration(p.TimeWindow)
+		w, err := time.ParseDuration(p.TimeWindow)
 		if err != nil {
 			return nil, err
 		}
+
+		if w < metricTimeWindowLowerBound {
+			return nil, errors.New("metrics time window needs to be at least 15s")
+		}
+
 		window = p.TimeWindow
 	}