+++ test/install_test.go
@@ -1,7 +1,6 @@
 package test
 
 import (
-	"encoding/json"
 	"fmt"
 	"os"
 	"regexp"
@@ -136,6 +135,10 @@ func TestVersionPreInstall(t *testing.T) {
 }
 
 func TestCheckPreInstall(t *testing.T) {
+	if TestHelper.ExternalIssuer() {
+		t.Skip("Skipping pre-install check for external issuer test")
+	}
+
 	if TestHelper.UpgradeFromVersion() != "" {
 		t.Skip("Skipping pre-install check for upgrade test")
 	}
@@ -213,8 +216,53 @@ func TestInstallOrUpgradeCli(t *testing.T) {
 		}
 	)
 
-	if TestHelper.GetClusterDomain() != "" {
+	if TestHelper.GetClusterDomain() != "cluster.local" {
 		args = append(args, "--cluster-domain", TestHelper.GetClusterDomain())
+	}
+
+	if TestHelper.ExternalIssuer() {
+
+		// short cert lifetime to put some pressure on the CSR request, response code path
+		args = append(args, "--identity-issuance-lifetime=15s", "--identity-external-issuer=true")
+
+		err := TestHelper.CreateControlPlaneNamespaceIfNotExists(TestHelper.GetLinkerdNamespace())
+		if err != nil {
+			t.Fatalf("failed to create %s namespace: %s", TestHelper.GetLinkerdNamespace(), err)
+		}
+
+		identity := fmt.Sprintf("identity.%s.%s", TestHelper.GetLinkerdNamespace(), TestHelper.GetClusterDomain())
+
+		root, err := tls.GenerateRootCAWithDefaults(identity)
+		if err != nil {
+			t.Fatal(err)
+		}
+
+		// instead of passing the roots and key around we generate
+		// two secrets here. The second one will be used in the
+		// external_issuer_test to update the first one and trigger
+		// cert rotation in the identity service. That allows us
+		// to generated the certs on the fly and use custom domain.
+
+		if err = TestHelper.CreateTLSSecret(
+			k8s.IdentityIssuerSecretName,
+			root.Cred.Crt.EncodeCertificatePEM(),
+			root.Cred.Crt.EncodeCertificatePEM(),
+			root.Cred.EncodePrivateKeyPEM()); err != nil {
+			t.Fatal(err)
+		}
+
+		crt2, err := root.GenerateEndEntityCred(identity)
+		if err != nil {
+			t.Fatal(err)
+		}
+
+		if err = TestHelper.CreateTLSSecret(
+			k8s.IdentityIssuerSecretName+"-new",
+			root.Cred.Crt.EncodeCertificatePEM(),
+			crt2.EncodeCertificatePEM(),
+			crt2.EncodePrivateKeyPEM()); err != nil {
+			t.Fatal(err)
+		}
 	}
 
 	if TestHelper.UpgradeFromVersion() != "" {
@@ -237,9 +285,9 @@ func TestInstallOrUpgradeCli(t *testing.T) {
 	}
 
 	exec := append([]string{cmd}, args...)
-	out, _, err := TestHelper.LinkerdRun(exec...)
+	out, stderr, err := TestHelper.LinkerdRun(exec...)
 	if err != nil {
-		t.Fatalf("linkerd install command failed\n%s", out)
+		t.Fatalf("linkerd install command failed: \n%s\n%s\n%s", out, stderr, out)
 	}
 
 	// test `linkerd upgrade --from-manifests`
@@ -468,7 +516,7 @@ func TestInject(t *testing.T) {
 
 			prefixedNs := TestHelper.GetTestNamespace(tc.ns)
 
-			err := TestHelper.CreateNamespaceIfNotExists(prefixedNs, tc.annotations)
+			err := TestHelper.CreateDataPlaneNamespaceIfNotExists(prefixedNs, tc.annotations)
 			if err != nil {
 				t.Fatalf("failed to create %s namespace: %s", prefixedNs, err)
 			}
@@ -654,22 +702,14 @@ func TestEvents(t *testing.T) {
 	if err != nil {
 		t.Errorf("kubectl get events command failed with %s\n%s", err, out)
 	}
-	var list corev1.List
-	if err := json.Unmarshal([]byte(out), &list); err != nil {
-		t.Errorf("Error unmarshaling list from `kubectl get events`: %s", err)
-	}
 
-	if len(list.Items) == 0 {
-		t.Error("No events found")
+	events, err := testutil.ParseEvents(out)
+	if err != nil {
+		t.Fatal(err)
 	}
 
 	var unknownEvents []string
-	for _, i := range list.Items {
-		var e corev1.Event
-		if err := json.Unmarshal(i.Raw, &e); err != nil {
-			t.Errorf("Error unmarshaling list event from `kubectl get events`: %s", err)
-		}
-
+	for _, e := range events {
 		if e.Type == corev1.EventTypeNormal {
 			continue
 		}